feat: refactor home page
to do with #446
